==> notebooks/01_get_facility.ipynb <==
import pandas as pd


df = pd.read_csv("../data/raw/srv_ifmd_all-facilities.csv")

df.head()

# extract columns about facilities
df = df.loc[:,['Facility ID','Facility Name','Street Name','Suburb/Town','Pcode','Latitude',
                'Longitude','Field/Surface Type','FullAddress','MelwaysVicRoadsRef']]

df.head()
df.shape

# remove duplications
new = df.drop_duplicates(subset=['FullAddress','Latitude','Longitude'], keep='first')

new.shape

# distinguish different types of facility
print(new['Facility Name'].str.contains('park|reserve', case=False))

# add each type of facility to the dataframe
park_reserve = new[new['Facility Name'].str.contains('park|reserve', case=False)]

gym_fitness = new[new['Facility Name'].str.contains('gym|fitness', case=False)]

swimming_sport = new[new['Facility Name'].str.contains('swimming', case=False)]

school = new[new['Facility Name'].str.contains('college|university', case=False)]

# reset index and put each row in order
park_reserve.reset_index(drop=True, inplace=True)

gym_fitness.reset_index(drop=True, inplace=True)

swimming_sport.reset_index(drop=True, inplace=True)

school.reset_index(drop=True, inplace=True)

park_reserve.head()

park_reserve.to_csv("../data/curated/park_reserve.csv")

gym_fitness.to_csv("../data/curated/gym_fitness.csv")

swimming_sport.to_csv("../data/curated/swimming.csv")

school.to_csv("../data/curated/school.csv")
