==> notebooks/02_income_clean.ipynb <==
import pandas as pd

income = pd.read_csv("../data/raw/income.csv")
income.head()

# extract victoria's data
income_new = income.iloc[584:1047]

income_new.head()

# clean noisy data
income_new.dropna(how='all', axis=1, inplace=True)
income_new.head()

new = income_new.iloc[:, 0:7]

# arrange data by SA2 name for future matching with domain data
new.set_axis(["SA2", "SA2 NAME", "2014-15", "2015-16", "2016-17", "2017-18", "2018-19"],axis = 1, inplace = True)
new.head()

list_income = ["SA2 NAME", "2014-15", "2015-16", "2016-17", "2017-18", "2018-19"]
# recreate dataframe with only useful columns
df = new.loc[:, list_income]
df = df.rename(columns={'SA2 NAME':'suburb', '2014-15':'2014income', \
                                '2015-16':'2015income', '2016-17':'2016income', \
                                '2017-18':'2017income', '2018-19':'2018income'})

df.reset_index(drop=True, inplace=True)

df.to_csv("../data/curated/income_person.csv")
