==> notebooks/04_domain_cleaned.ipynb <==
import pandas as pd
  
df_apa = pd.DataFrame(pd.read_csv("../data/raw/domain_apartment.csv"))
df_hou = pd.DataFrame(pd.read_csv("../data/raw/domain_house.csv"))


# Removal of parking rental
df_apa = df_apa[(df_apa['beds'] != 0) & (df_apa['baths'] != 0)]
df_hou = df_hou[(df_hou['beds'] != 0) & (df_hou['baths'] != 0)]

# remove comma in price
df_apa["price"] = [df_apa["price"][i].replace(',', '') for i in range(len(df_apa))]
df_hou["price"] = [df_hou["price"][i].replace(',', '') for i in range(len(df_hou))]

# find the pw price
df_apa["price"] = df_apa.price.astype(str).str.slice(1,6)
df_hou["price"] = df_hou.price.astype(str).str.slice(1,6)

# Only keep the digits
df_apa["price"] = df_apa["price"].str.extract('(\d+)')
df_hou["price"] = df_hou["price"].str.extract('(\d+)')

# remove other types
apartment = df_apa.query('propertyTypeFormatted==["Apartment / Unit / Flat"]')
house = df_hou.query('propertyTypeFormatted==["House"]')


# remove missing values
cleaned_apa = apartment.dropna()
cleaned_hou = house.dropna()

# save 
cleaned_apa.to_csv('../data/curated/apartment.csv', index=False)
cleaned_hou.to_csv('../data/curated/house.csv', index=False)
