==> notebooks/05_geocode.ipynb <==
pip install geopy

import geopy
import geopandas
import pandas as pd

from opencage.geocoder import OpenCageGeocode

key = "YOUR_OWN_KEY"
geocoder = OpenCageGeocode(key) 

df1 = pd.read_csv("../data/curated/house.csv")

df2 = pd.read_csv("../data/curated/apartment.csv")

df3 = pd.read_csv("../data/curated/park_reserve.csv")

addresses = df1["address"].values.tolist()

latitudes = []
longitudes = []
# calcualate longitudes and latitudes of house
for address in addresses:
    result = geocoder.geocode(address, no_annotations="1")

    if result and len(result):
        longitude = result[0]['geometry']['lng']
        latitude = result[0]['geometry']['lat']
    else:
        longitude = "N/A"
        latitude = "N/A"

    latitudes.append(latitude)
    longitudes.append(longitude)

df1["latitudes"] = latitudes
df1["longitudes"] = longitudes

df1.to_csv("../data/curated/house_geocoded.csv")

addresses2 = df2["address"].values.tolist()

latitudes2 = []
longitudes2 = []

for address in addresses2:
    result = geocoder.geocode(address, no_annotations="1")

    if result and len(result):
        longitude = result[0]['geometry']['lng']
        latitude = result[0]['geometry']['lat']
    else:
        longitude = "N/A"
        latitude = "N/A"

    latitudes2.append(latitude)
    longitudes2.append(longitude)

df2["latitudes"] = latitudes2
df2["longitudes"] = longitudes2

df2.to_csv("../data/curated/apartment_geocoded.csv")

addresses = df3["address"].values.tolist()

latitudes = []
longitudes = []
# calcualate longitudes and latitudes of park
for address in addresses:
    result = geocoder.geocode(address, no_annotations="1")

    if result and len(result):
        longitude = result[0]['geometry']['lng']
        latitude = result[0]['geometry']['lat']
    else:
        longitude = "N/A"
        latitude = "N/A"

    latitudes.append(latitude)
    longitudes.append(longitude)

df3["latitudes"] = latitudes
df3["longitudes"] = longitudes

df3.to_csv("../data/curated/park_geocoded.csv")
