==> notebooks/06_geocode_clean.ipynb <==
import pandas as pd

df1 = pd.read_csv('../data/curated/apartment_geocoded.csv')
df2 = pd.read_csv('../data/curated/house_geocoded.csv')

df1.head()

#apartment
df_clear1 = df1.drop(df1[df1['latitudes']>0].index)

#house
df_clear2 = df2.drop(df2[df2['latitudes']>0].index)

df_clear1

df_clear1.reset_index(drop=True, inplace=True)
df_clear2.reset_index(drop=True, inplace=True)

df_clear1["coordinates"] = [[df_clear1["longitudes"][i], df_clear1["latitudes"][i]] for i in range(len(df_clear1))]

df_clear1.head()

df_clear2["coordinates"] = [[df_clear2["longitudes"][i], df_clear2["latitudes"][i]] for i in range(len(df_clear2))]

df_clear1.to_csv('../data/curated/clean_apartment_geo.csv')
df_clear2.to_csv('../data/curated/clean_house_geo.csv')
