==> notebooks/07_shoppingmalls_srcape.ipynb <==
import re
import pandas as pd
import requests
from bs4 import BeautifulSoup


#get website data
def html_get():
    #Simulate browser request
    headers = {         
        "authority": "www.australia-shoppings.com",
        "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
        "accept-language": "zh-CN,zh;q=0.9,ja;q=0.8",
        "cache-control": "no-cache",
        "pragma": "no-cache",
        "sec-ch-ua": "^\\^Google",
        "sec-ch-ua-mobile": "?0",
        "sec-ch-ua-platform": "^\\^Windows^^",
        "sec-fetch-dest": "document",
        "sec-fetch-mode": "navigate",
        "sec-fetch-site": "none",
        "sec-fetch-user": "?1",
        "upgrade-insecure-requests": "1",
        "user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/105.0.0.0 Safari/537.36"
    }

    #shopping mall website
    url = "https://www.australia-shoppings.com/malls-centres/victoria"  
    response = requests.get(url, headers=headers)  

    #change to text type
    return response.text   

#analysis data and prepare to save
dic_list=[]   
def data_analysis(html):
    #Convert to bs4 format 
    soup = BeautifulSoup(html, "html.parser")  

    ul=soup.find('ul',attrs='malls-list')    

    li=ul.find_all('li')       
    for item in li:        
        #get address
        address=item.find('h3').find('a').string  
        

        all=item.find('p')
        #regular expression
        all=re.sub('<.*?>','',str(all))  
        all=all.split(',')              
        

        area  = all[0:2]              
        area=','.join(area)           
      
       # get longitude and latitude
        Longitude = str(all[-1]).split("|")[0]          
        Latitude = all[-2].replace('GPS:',"")
       
        dic={                       
            "Full Name":address,
            "Address":area,
            "Longitude": Longitude,
            "Latitude": Latitude
        }
        
        dic_list.append(dic)         



html=html_get()             #get website data
data_analysis(html)         #analysis data
df=pd.DataFrame(dic_list)   #change data type to df
df.to_csv("../data/raw/shoppingmall_raw.csv")    #save as csv

df
