==> notebooks/09_Data preprocessing.ipynb <==
import pandas as pd
from pandas import Series,DataFrame

# read datafiles
dfapart = pd.read_csv("../data/curated/apartment_facilities.csv")
dfhouse = pd.read_csv('../data/curated/house_facilities.csv')
income = pd.read_csv('../data/curated/income_person.csv')

population = pd.read_csv("../data/raw/population_2021.csv")
population = population.rename(columns={'sa2_name_2016':'suburb'})
# convert all string into upper form
income['suburb'] = income['suburb'].str.upper()
population['suburb'] = population['suburb'].str.upper()

houseprice = pd.read_csv('../data/raw/Suburb_Unit_price.csv')
houseprice = houseprice.rename(columns={'Unnamed: 0':'suburb'})

# find the suburb of each address
dfapart["suburb"] = dfapart["address"].str.extract("(.+?)\d")
dfapart["suburb"] = dfapart["suburb"].str.strip().str.upper()

dfhouse["suburb"] = dfhouse["address"].str.extract("(.+?)\d")
dfhouse["suburb"] = dfhouse["suburb"].str.strip().str.upper()

apa = dfapart.loc[:,["price", "suburb", "beds", "baths", \
                 "parking", "latitudes", "longitudes", \
                 "school", "train_station", "fitness_gym", \
                 "hospital", "shopping_center", "park"]]
house = dfhouse.loc[:,["price", "suburb", "beds", "baths", \
                 "parking", "latitudes", "longitudes", \
                 "school", "train_station", "fitness_gym", \
                 "hospital", "shopping_center", "park"]]

# convert to string type
house["suburb"] = house["suburb"].apply(str)
apa["suburb"] = apa["suburb"].apply(str)

list_income = ["suburb", "2014income", "2015income", "2016income", "2017income", "2018income"]
income = income.loc[:, list_income]

# create dataframe of useful information for population
pop_list = ['suburb', 'erp_2016', 'erp_2017', 'erp_2018', 'erp_2019', 'erp_2020',\
            'natural_increase_2016_17', 'net_internal_migration_2016_17','net_overseas_migration_2016_17', \
                 'natural_increase_2017_18', 'net_internal_migration_2017_18', 'net_overseas_migration_2017_18',\
       'natural_increase_2018_19', 'net_internal_migration_2018_19', 'net_overseas_migration_2018_19',\
            'natural_increase_2019_20', 'net_internal_migration_2019_20', 'net_overseas_migration_2019_20', \
            'natural_increase_2020_21', 'net_internal_migration_2020_21', 'net_overseas_migration_2020_21', ]
population = population.loc[:,pop_list]

population.to_csv('../data/curated/population16-20.csv')

houseprice.to_csv('../data/curated/houseprice.csv')
