==> notebooks/10_pop_pred.ipynb <==
### Data cleaning


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd
from sklearn.linear_model import LinearRegression

df = pd.read_csv("../data/curated/population16-20.csv")
df.shape

df.head(10)

# check for missing value
df.isnull().sum()

df.set_index(["suburb"], inplace=True)
df

df = df.apply(pd.to_numeric)

# check for unreasonable value
df = df[(df.erp_2016> 16) & (df.erp_2017> 16) & (df.erp_2018> 16) &(df.erp_2019> 16)
        &(df.erp_2020> 16)]
df.shape

#seperate to 4 dfs
df_erp = df[['erp_2016','erp_2017','erp_2018','erp_2019','erp_2020']]
df_nc = df[['natural_increase_2016_17','natural_increase_2017_18','natural_increase_2018_19',
            'natural_increase_2019_20','natural_increase_2020_21']]
df_iter_mig = df[['net_internal_migration_2016_17','net_internal_migration_2017_18','net_internal_migration_2018_19',
            'net_internal_migration_2019_20','net_internal_migration_2020_21']]
df_os_mig = df[['net_overseas_migration_2016_17','net_overseas_migration_2017_18','net_overseas_migration_2018_19',
            'net_overseas_migration_2019_20','net_overseas_migration_2020_21']]
df_os_mig

### Model: population forecasting(End of Sprint 4)


df_erp

#for saving predicted erp
l21_erp = []
l22_erp= []
l23_erp= []
l24_erp = []
l25_erp = []

# recreate dataframe with only year and population
for i in range(len(df_erp)):
    df1 = pd.DataFrame(data=[['2016',df_erp.iloc[i,0]],
                    ['2017',df_erp.iloc[i,1]],
                    ['2018',df_erp.iloc[i,2]],
                    ['2019', df_erp.iloc[i,3]],
                    ['2020',df_erp.iloc[i,4]]],
             columns = ['year','erp'],
             )
    y = df1['erp']
    x = df1['year']
    x = x.values.reshape(-1,1)
    y = y.values.reshape(-1,1)
    model = LinearRegression()
    model.fit(x, y)
    
    # predict populations for 3 years
    predict_y = model.predict(x)
    X2 = [[2021],[2022],[2023],[2024],[2025]]
    Y2 = model.predict(X2)
    l21_erp.append(Y2[0][0])
    l22_erp.append(Y2[1][0])
    l23_erp.append(Y2[2][0])
    l24_erp.append(Y2[3][0])
    l25_erp.append(Y2[4][0])

len(l21_erp)

import os
os.mkdir(r'../data/curated/predictions')

# append to new dataframe
df_erp["erp_2021"] = l21_erp
df_erp["erp_2022"] = l22_erp
df_erp["erp_2023"] = l23_erp
df_erp["erp_2024"] = l24_erp
df_erp["erp_2025"] = l25_erp

df_erp.to_csv('../data/curated/predictions/erp.csv')

#for saving predicted natural increase
l21_nc = []
l22_nc = []
l23_nc= []
l24_nc = []
l25_nc = []

# same as before
for i in range(len(df_nc)):
    df1 = pd.DataFrame(data=[['2016',df_nc.iloc[i,0]],
                    ['2017',df_nc.iloc[i,1]],
                    ['2018',df_nc.iloc[i,2]],
                    ['2019', df_nc.iloc[i,3]],
                    ['2020',df_nc.iloc[i,4]]],
             columns = ['year','natural increase'],
             )
    y = df1['natural increase']
    x = df1['year']
    x = x.values.reshape(-1,1)
    y = y.values.reshape(-1,1)
    model = LinearRegression()
    model.fit(x, y)
    
    # predict natural increase for 3 years
    predict_y = model.predict(x)
    X2 = [[2021],[2022],[2023],[2024],[2025]]
    Y2 = model.predict(X2)
    l21_nc.append(Y2[0][0])
    l22_nc.append(Y2[1][0])
    l23_nc.append(Y2[2][0])
    l24_nc.append(Y2[3][0])
    l25_nc.append(Y2[4][0])
    


df_nc["natural_increase_2020_21"] = l21_nc
df_nc["natural_increase_2021_22"] = l22_nc
df_nc["natural_increase_2022_23"] = l23_nc
df_nc['natural_increase_2023_24'] = l24_nc
df_nc["natural_increase_2024_25"] = l25_nc
df_nc.to_csv('../data/curated/predictions/natural_increase.csv')

#for saving predicted iternal migration
l21_iter_mig = []
l22_iter_mig= []
l23_iter_mig= []
l24_iter_mig = []
l25_iter_mig = []

df_iter_mig

for i in range(len(df_iter_mig)):
    df1 = pd.DataFrame(data=[['2016',df_iter_mig.iloc[i,0]],
                    ['2017',df_iter_mig.iloc[i,1]],
                    ['2018',df_iter_mig.iloc[i,2]],
                    ['2019', df_iter_mig.iloc[i,3]],
                    ['2020',df_iter_mig.iloc[i,4]]],
             columns = ['year','iter_mig'],
             )
    y = df1['iter_mig']
    x = df1['year']
    x = x.values.reshape(-1,1)
    y = y.values.reshape(-1,1)
    model = LinearRegression()
    model.fit(x, y)
    
    # predict iternal migration for 3 years
    predict_y = model.predict(x)
    X2 = [[2021],[2022],[2023],[2024],[2025]]
    Y2 = model.predict(X2)
    l21_iter_mig.append(Y2[0][0])
    l22_iter_mig.append(Y2[1][0])
    l23_iter_mig.append(Y2[2][0])
    l24_iter_mig.append(Y2[3][0])
    l25_iter_mig.append(Y2[4][0])

df_iter_mig["net_internal_migration_2020_21"] = l21_iter_mig
df_iter_mig["net_internal_migration_2021_22"] = l22_iter_mig
df_iter_mig["net_internal_migration_2022_23"] = l23_iter_mig
df_iter_mig['net_internal_migration_2023_24'] = l24_iter_mig
df_iter_mig["net_internal_migration_2024_25"] = l25_iter_mig
df_iter_mig.to_csv('../data/curated/predictions/net_internal_migration.csv')

#for saving predicted overseas migration
l21_os_mig = []
l22_os_mig= []
l23_os_mig= []
l24_os_mig = []
l25_os_mig = []

for i in range(len(df_os_mig)):
    df1 = pd.DataFrame(data=[['2016',df_os_mig.iloc[i,0]],
                    ['2017',df_os_mig.iloc[i,1]],
                    ['2018',df_os_mig.iloc[i,2]],
                    ['2019', df_os_mig.iloc[i,3]],
                    ['2020',df_os_mig.iloc[i,4]]],
             columns = ['year','os_mig'],
             )
    y = df1['os_mig']
    x = df1['year']
    x = x.values.reshape(-1,1)
    y = y.values.reshape(-1,1)
    model = LinearRegression()
    model.fit(x, y)
    
    # predict overseas-migration for 3-years
    predict_y = model.predict(x)
    X2 = [[2021],[2022],[2023],[2024],[2025]]
    Y2 = model.predict(X2)
    l21_os_mig.append(Y2[0][0])
    l22_os_mig.append(Y2[1][0])
    l23_os_mig.append(Y2[2][0])
    l24_os_mig.append(Y2[3][0])
    l25_os_mig.append(Y2[4][0])

df_os_mig["net_overseas_migration_2020_21"] = l21_os_mig
df_os_mig["net_overseas_migration_2021_22"] = l22_os_mig
df_os_mig["net_overseas_migration_2022_23"] = l23_os_mig
df_os_mig['net_overseas_migration_2023_24'] = l24_os_mig
df_os_mig["net_overseas_migration_2024_25"] = l25_os_mig
df_os_mig.to_csv('../data/curated/predictions/net_overseas_migration.csv')
