==> notebooks/12_merge_datasets.ipynb <==
import pandas as pd
from pandas import Series,DataFrame

# read datafiles
dfapart = pd.read_csv("../data/curated/apartment_facilities.csv")
dfhouse = pd.read_csv('../data/curated/house_facilities.csv')
income = pd.read_csv('../data/curated/predictions/pred_income.csv')
erp = pd.read_csv("../data/curated/predictions/erp.csv")
internal = pd.read_csv("../data/curated/predictions/net_internal_migration.csv")
oversea = pd.read_csv("../data/curated/predictions/net_overseas_migration.csv")
natural_inc = pd.read_csv("../data/curated/predictions/natural_increase.csv")
houseprice = pd.read_csv("../data/curated/predictions/pred_house_price.csv")

erp.columns

# select useful columns
income = income.loc[:,['suburb','2022income', '2023income', '2024income', '2025income']]
erp = erp.loc[:,['suburb', 'erp_2022', 'erp_2023', 'erp_2024', 'erp_2025']]
internal = internal.loc[:,['suburb', 'net_internal_migration_2021_22', 'net_internal_migration_2022_23',\
                           'net_internal_migration_2023_24', 'net_internal_migration_2024_25']]
oversea = oversea.loc[:,['suburb', 'net_overseas_migration_2021_22', 'net_overseas_migration_2022_23', \
                         'net_overseas_migration_2023_24', 'net_overseas_migration_2024_25']]
natural_inc = natural_inc.loc[:,['suburb', 'natural_increase_2021_22','natural_increase_2022_23', \
                                 'natural_increase_2023_24','natural_increase_2024_25']]
houseprice = houseprice.loc[:,['suburb', '2022house_price', '2023house_price', '2024house_price', '2025house_price']]

# separate years
for i in range(1, len(erp.columns)):
    erp1 = erp.iloc[:,[0,i]]
    internal1 = internal.iloc[:,[0,i]]
    oversea1 = oversea.iloc[:,[0,i]]
    natural_inc1 = natural_inc.iloc[:,[0,i]]
    df_merged = pd.merge(erp1, internal1, on=['suburb'])
    df_merged2 = pd.merge(df_merged, oversea1, on=['suburb'])
    df_merged3 = pd.merge(df_merged2, natural_inc1, on=['suburb'])
    df_merged3.to_csv(f"../data/curated/predictions/{i}population_prediction.csv")

pop1 = pd.read_csv("../data/curated/predictions/1population_prediction.csv")
pop2 = pd.read_csv("../data/curated/predictions/2population_prediction.csv")
pop3 = pd.read_csv("../data/curated/predictions/3population_prediction.csv")
pop4 = pd.read_csv("../data/curated/predictions/4population_prediction.csv")

# find the suburb of each address
dfapart["suburb"] = dfapart["address"].str.extract("(.+?)\d")
dfapart["suburb"] = dfapart["suburb"].str.strip().str.upper()

dfhouse["suburb"] = dfhouse["address"].str.extract("(.+?)\d")
dfhouse["suburb"] = dfhouse["suburb"].str.strip().str.upper()

apa = dfapart.loc[:,["price", "suburb", "beds", "baths", \
                 "parking", "latitudes", "longitudes", \
                 "school", "train_station", "fitness_gym", \
                 "hospital", "shopping_center", "park"]]
house = dfhouse.loc[:,["price", "suburb", "beds", "baths", \
                 "parking", "latitudes", "longitudes", \
                 "school", "train_station", "fitness_gym", \
                 "hospital", "shopping_center", "park"]]

house["suburb"] = house["suburb"].apply(str)
apa["suburb"] = apa["suburb"].apply(str)

# list SA2 names
sub = pop1["suburb"]
sub = sub.dropna(how ="any")
sub = sub.str.upper().to_list()
income["suburb"] = income["suburb"].str.upper()

# match the corresponding area of houseprice
from difflib import get_close_matches
for i in range(len(houseprice)):
    area = houseprice["suburb"][i]
    s = get_close_matches(area, sub, n=1, cutoff=0.8)
    houseprice["suburb"][i] = s
    if len(houseprice["suburb"][i])==0:
        houseprice["suburb"][i] = ''
    else:
        houseprice["suburb"][i] = houseprice["suburb"][i][0]    
      

from difflib import get_close_matches
# match the corresponding area of income
for i in range(len(income)):
    area = income["suburb"][i]
    s = get_close_matches(area, sub, n=1, cutoff=0.7)
    income["suburb"][i] = s
    if len(income["suburb"][i])==0:
        income["suburb"][i] = ''
    else:
        income["suburb"][i] = income["suburb"][i][0]    
      

# match the corresponding area of house
from difflib import get_close_matches
for i in range(len(house)):
    area = house["suburb"][i]
    s = get_close_matches(area, sub, n=1)
    house["suburb"][i] = s
    if len(house["suburb"][i])==0:
        house["suburb"][i] = ''
    else:
        house["suburb"][i] = house["suburb"][i][0]    
        

# match the corresponding area of apartment
for i in range(len(apa)):
    area = apa["suburb"][i]
    s = get_close_matches(area, sub, n=1)
    apa["suburb"][i] = s
    if len(apa["suburb"][i])==0:
        apa["suburb"][i] = ''
    else:
        apa["suburb"][i] = apa["suburb"][i][0]   
        

list_pop = [pop1, pop2, pop3, pop4]

# match population and population and income
for i in range(1, len(erp.columns)):
    income1 = income.iloc[:,[0,i]]
    houseprice1 = houseprice.iloc[:,[0,i]]
    pop = list_pop[i-1]
    df_merged = pd.merge(income1, houseprice1, on=['suburb'])
    df_merged2 = pd.merge(df_merged, pop, on=['suburb'])
    matchhouse = pd.merge(house, df_merged2, how = 'inner', on = ['suburb'])
    matchapart = pd.merge(apa, df_merged2, how = 'inner', on = ['suburb'])
    matchhouse = matchhouse.drop_duplicates()
    matchapart = matchapart.drop_duplicates()
    print(f'{i+21}')
    print(len(matchhouse))
    print(len(matchapart))
    matchapart.to_csv(f'../data/curated/match_apartment{i+21}.csv')
    matchhouse.to_csv(f'../data/curated/match_house{i+21}.csv')
