==> notebooks/18_geomap.ipynb <==
import folium
import geopandas as gpd
import pandas as pd
import numpy as np
import json
import requests
import zipfile
from io import BytesIO

# download the compressed shapefile used to match location info
print('Downloading started')
url_shape = "https://data.gov.au/data/dataset/af33dd8c-0534-4e18-9245-fc64440f742e/resource/02dd5644-7542-4e95-ab71-1707010b99a6/download/vic_localities.zip"
req = requests.get(url_shape)
file = "../data/raw/vic_localities.zip"

# Writing the file to the local file system
with open(file,'wb') as output_file:
    output_file.write(req.content)
print('Zip: Shapefile Downloading Completed')

# extracting the zip file contents
zipfile= zipfile.ZipFile(BytesIO(req.content))
zipfile.extractall('../data/raw/vic_localities')

zones = gpd.read_file("../data/raw/vic_localities/vic_localities.shp")
zones.head()

# express location in latitude and longitude and unify the format
zones['geometry'] = zones['geometry'].to_crs("+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs")
zones.rename(columns={'LOC_NAME':'suburb'},inplace = True)
zones['suburb'] = zones['suburb'].str.upper()

#convert to json
geoJSON = zones[['suburb', 'geometry']].to_json()
print(geoJSON[:300])

map_apa = folium.Map(location=[ -37.8136, 144.9631], tiles="Stamen Terrain", zoom_start=10)


# read data
apa = pd.read_csv('../data/curated/match_apartment22_s.csv')
house = pd.read_csv('../data/curated/match_house22_s.csv')

#calculate average house and apartment price in each suburb
agg_df = apa[['suburb', 'price']].groupby('suburb').mean()
agg_df = agg_df.nlargest(50,'price')
agg_df['suburb'] = agg_df.index
agg_df.head()

house_df = house[['suburb', 'price']].groupby('suburb').mean()
house_df = house_df.nlargest(50,'price')
house_df['suburb'] = house_df.index
house_df.head()

agg_df['price'].describe()


c = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=agg_df, # data source
    columns=['suburb','price'], # the columns required
    key_on='properties.suburb', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color = 'silver',
    highlight=True,
    line_weight= 0.5,
    legend_name='mean price'
)

c.add_to(map_apa)


map_apa.save('../plots/apa_priceMap.html')

map_house = folium.Map(location=[ -37.8136, 144.963], tiles="Stamen Terrain", zoom_start=10)

h = folium.Choropleth(
    geo_data=geoJSON, # geoJSON 
    name='choropleth', # name of plot
    data=house_df, # data source
    columns=['suburb','price'], # the columns required
    key_on='properties.suburb', # this is from the geoJSON's properties
    fill_color='YlOrRd', # color scheme
    nan_fill_color = 'silver',
    highlight=True,
    line_weight= 0.5,
    legend_name='mean price'
)

h.add_to(map_house)
map_house.save('../plots/house_priceMap.html')
